+++ src/Libraries/SmartStore.Services/DataExchange/Export/IDataExporter.cs
@@ -17,20 +17,20 @@ namespace SmartStore.Services.DataExchange.Export
         DataExportPreviewResult Preview(DataExportRequest request, int pageIndex);
 
 		/// <summary>
-		/// Creates a product export context for fast retrieval (eager loading) of product navigation properties
+		/// Creates a product export context for fast retrieval (eager loading) of product navigation properties.
 		/// </summary>
 		/// <param name="products">Products. <c>null</c> to lazy load data if required.</param>
 		/// <param name="customer">Customer, <c>null</c> to use current customer.</param>
 		/// <param name="storeId">Store identifier, <c>null</c> to use current store.</param>
 		/// <param name="maxPicturesPerProduct">Pictures per product, <c>null</c> to load all pictures per product.</param>
-		/// <param name="showHidden">A value indicating whether to show hidden records</param>
+		/// <param name="includeHidden">A value indicating whether to include hidden records.</param>
 		/// <returns>Product export context</returns>
 		ProductExportContext CreateProductExportContext(
 			IEnumerable<Product> products = null,
 			Customer customer = null,
 			int? storeId = null,
 			int? maxPicturesPerProduct = null,
-			bool showHidden = true);
+			bool includeHidden = true);
 	}