+++ src/Presentation/SmartStore.Web/Controllers/ProductController.cs
@@ -643,14 +643,15 @@ namespace SmartStore.Web.Controllers
 			if (product == null)
 			{
 				throw new ArgumentException(T("Products.NotFound", productId));
-			}				
+			}
 
-			var cacheKey = string.Format(ModelCacheEventConsumer.PRODUCTTAG_BY_PRODUCT_MODEL_KEY, product.Id, _services.WorkContext.WorkingLanguage.Id, _services.StoreContext.CurrentStore.Id);
+            var store = _services.StoreContext.CurrentStore;
+
+            var cacheKey = string.Format(ModelCacheEventConsumer.PRODUCTTAG_BY_PRODUCT_MODEL_KEY, product.Id, _services.WorkContext.WorkingLanguage.Id, store.Id);
 			var cacheModel = _services.Cache.Get(cacheKey, () =>
 			{
 				var model = product.ProductTags
-					// Filter by store
-					.Where(x => _productTagService.GetProductCount(x.Id, _services.StoreContext.CurrentStore.Id) > 0)
+					.Where(x => x.Published && _productTagService.GetProductCount(x.Id, store.Id) > 0)
 					.Select(x =>
 					{
 						var ptModel = new ProductTagModel
@@ -658,7 +659,7 @@ namespace SmartStore.Web.Controllers
 							Id = x.Id,
 							Name = x.GetLocalized(y => y.Name),
 							SeName = x.GetSeName(),
-							ProductCount = _productTagService.GetProductCount(x.Id, _services.StoreContext.CurrentStore.Id)
+							ProductCount = _productTagService.GetProductCount(x.Id, store.Id)
 						};
 						return ptModel;
 					})